refactor(chat): move error UI into dedicated page component and re-export
- Extract full error dialog and local storage clearing logic from error.tsx to error/page.tsx
- Update error.tsx to re-export the new default component
- Clean up comments in proxy middleware configuration
